Copy ManagementController into the separate applications
