+++ catalogue_pipeline/id_minter/src/main/scala/uk/ac/wellcome/platform/idminter/Server.scala
@@ -8,12 +8,16 @@ import com.twitter.finatra.http.filters.{
   TraceIdMDCFilter
 }
 import com.twitter.finatra.http.routing.HttpRouter
-import uk.ac.wellcome.finatra.controllers.ManagementController
 import uk.ac.wellcome.finatra.modules._
 import uk.ac.wellcome.messaging.message.MessageConfigModule
+import uk.ac.wellcome.monitoring.CloudWatchClientModule
+import uk.ac.wellcome.monitoring.MetricsSenderModule
 import uk.ac.wellcome.messaging.sns.SNSClientModule
-import uk.ac.wellcome.messaging.sqs.{SQSClientModule, SQSReaderModule}
-import uk.ac.wellcome.monitoring.{CloudWatchClientModule, MetricsSenderModule}
+import uk.ac.wellcome.messaging.sqs.{
+  SQSClientModule,
+  SQSReaderModule
+}
+import uk.ac.wellcome.platform.idminter.controllers.ManagementController
 import uk.ac.wellcome.platform.idminter.modules._
 import uk.ac.wellcome.storage.s3.S3ClientModule
