+++ catalogue_pipeline/id_minter/src/main/scala/uk/ac/wellcome/platform/idminter/controllers/ManagementController.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.api.controllers
+package uk.ac.wellcome.platform.idminter.controllers
 
 import javax.inject.{Inject, Singleton}
