+++ catalogue_pipeline/ingestor/src/main/scala/uk/ac/wellcome/platform/ingestor/controllers/ManagementController.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.finatra.controllers
+package uk.ac.wellcome.platform.ingestor.controllers
 
 import javax.inject.{Inject, Singleton}
