+++ catalogue_pipeline/recorder/src/main/scala/uk/ac/wellcome/platform/recorder/Server.scala
@@ -4,11 +4,11 @@ import com.twitter.finagle.http.{Request, Response}
 import com.twitter.finatra.http.HttpServer
 import com.twitter.finatra.http.filters.{CommonFilters, LoggingMDCFilter, TraceIdMDCFilter}
 import com.twitter.finatra.http.routing.HttpRouter
-import uk.ac.wellcome.finatra.controllers.ManagementController
 import uk.ac.wellcome.finatra.modules._
 import uk.ac.wellcome.messaging.message.MessageConfigModule
 import uk.ac.wellcome.monitoring.MetricsSenderModule
 import uk.ac.wellcome.messaging.sqs.{SQSClientModule, SQSConfigModule}
+import uk.ac.wellcome.platform.recorder.controllers.ManagementController
 import uk.ac.wellcome.platform.recorder.modules.{RecorderModule, RecorderWorkEntryKeyPrefixGeneratorModule}
 import uk.ac.wellcome.storage.dynamo.DynamoClientModule
 import uk.ac.wellcome.storage.s3.S3ClientModule
