+++ catalogue_pipeline/recorder/src/main/scala/uk/ac/wellcome/platform/recorder/controllers/ManagementController.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.api.controllers
+package uk.ac.wellcome.platform.recorder.controllers
 
 import javax.inject.{Inject, Singleton}
