+++ catalogue_pipeline/transformer/src/main/scala/uk/ac/wellcome/transformer/Server.scala
@@ -2,13 +2,8 @@ package uk.ac.wellcome.transformer
 
 import com.twitter.finagle.http.{Request, Response}
 import com.twitter.finatra.http.HttpServer
-import com.twitter.finatra.http.filters.{
-  CommonFilters,
-  LoggingMDCFilter,
-  TraceIdMDCFilter
-}
+import com.twitter.finatra.http.filters.{CommonFilters, LoggingMDCFilter, TraceIdMDCFilter}
 import com.twitter.finatra.http.routing.HttpRouter
-import uk.ac.wellcome.finatra.controllers.ManagementController
 import uk.ac.wellcome.finatra.modules._
 import uk.ac.wellcome.messaging.message.MessageConfigModule
 import uk.ac.wellcome.messaging.sns.SNSClientModule
@@ -16,6 +11,7 @@ import uk.ac.wellcome.messaging.sqs.{SQSClientModule, SQSConfigModule}
 import uk.ac.wellcome.monitoring.MetricsSenderModule
 import uk.ac.wellcome.storage.dynamo.DynamoConfigModule
 import uk.ac.wellcome.storage.s3.{S3ClientModule, S3ConfigModule}
+import uk.ac.wellcome.transformer.controllers.ManagementController
 import uk.ac.wellcome.transformer.modules.{TransformerWorkerModule, _}
 
 object ServerMain extends Server
