+++ data_api/snapshot_generator/src/main/scala/uk/ac/wellcome/platform/snapshot_generator/controllers/ManagementController.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.api.controllers
+package uk.ac.wellcome.platform.snapshot_generator.controllers
 
 import javax.inject.{Inject, Singleton}
