+++ sierra_adapter/sierra_bib_merger/src/main/scala/uk/ac/wellcome/platform/sierra_bib_merger/controllers/ManagementController.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.api.controllers
+package uk.ac.wellcome.platform.sierra_bib_merger.controllers
 
 import javax.inject.{Inject, Singleton}
