+++ sierra_adapter/sierra_reader/src/main/scala/uk/ac/wellcome/platform/sierra_reader/Server.scala
@@ -9,9 +9,9 @@ import com.twitter.finatra.http.filters.{
 }
 import com.twitter.finatra.http.routing.HttpRouter
 import uk.ac.wellcome.finatra.modules._
-import uk.ac.wellcome.finatra.controllers.ManagementController
 import uk.ac.wellcome.monitoring.MetricsSenderModule
 import uk.ac.wellcome.messaging.sqs.{SQSClientModule, SQSConfigModule}
+import uk.ac.wellcome.platform.sierra_reader.controllers.ManagementController
 import uk.ac.wellcome.platform.sierra_reader.modules.SierraReaderModule
 import uk.ac.wellcome.storage.s3.{S3ClientModule, S3ConfigModule}
