+++ sierra_adapter/sierra_reader/src/main/scala/uk/ac/wellcome/platform/sierra_reader/controllers/ManagementController.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.api.controllers
+package uk.ac.wellcome.platform.sierra_reader.controllers
 
 import javax.inject.{Inject, Singleton}
